+++ RELEASE_NOTES.md
@@ -4,4 +4,7 @@
 
 * Added support for deploying user-provided Helm charts
 * Added support for custom network configuration scripts
+
+### Image Definition
 * Removed the `embeddedArtifactRegistry/images/supplyChainKey` attribute
+* Changed `operatingSystem/users/sshKey` into `operatingSystem/users/sshKeys` and it is now a list instead of a single string